[mox] Publish separately from moxie crate.
It remains a dep of moxie-dom, which seems appropriate since it's "HTML".
