+++ Cargo.toml
@@ -17,23 +17,12 @@ edition = "2018"
 
 [dependencies]
 futures = "0.3.5"
+illicit = { path = "illicit", version = "1" }
 parking_lot = "0.11"
-proc-macro-hack = "0.5"
-proc-macro-nested = "0.1.3"
 scopeguard = "1"
+topo = { path = "topo", version = "0.10" }
 tracing = "^0.1"
 
-[dependencies.illicit]
-version = "1.0.0"
-path = "illicit"
-
-[dependencies.mox]
-version = "0.10.0-pre"
-path = "mox"
-
-[dependencies.topo]
-version = "0.10.0"
-path = "topo"
 
 [dev-dependencies]
 criterion = "0.3"
@@ -51,6 +40,7 @@ members = [
     "illicit",
     "illicit/macro",
     "mox",
+    "mox/impl",
     "topo",
     "topo/macro",
 ]
