+++ dom/examples/hacking/Cargo.toml
@@ -12,6 +12,7 @@ crate-type = [ "cdylib" ]
 
 [dependencies]
 console_log = "0.2.0"
+mox = { path = "../../../mox" }
 moxie-dom = { path = "../../" }
 topo = { path = "../../../topo" }
 tracing = { version = "^0.1", features = ["log"] }
