+++ dom/examples/ssr/Cargo.toml
@@ -11,6 +11,7 @@ augdom = { path = "../../augdom" }
 gotham = "0.4.0"
 gotham_derive = "0.4.0"
 hyper = "0.12"
+mox = { path = "../../../mox" }
 moxie = { path = "../../../" }
 serde = "1"
 serde_derive = "1"
