+++ dom/examples/todo/Cargo.toml
@@ -13,6 +13,7 @@ crate-type = [ "cdylib" ]
 [dependencies]
 console_log = "0.2.0"
 illicit = { path = "../../../illicit" }
+mox = { path = "../../../mox" }
 moxie-dom = { path = "../../" }
 topo = { path = "../../../topo" }
 tracing = { version = "^0.1", features = [ "log" ] }
