+++ dom/examples/todo/src/filter.rs
@@ -1,4 +1,5 @@
 use crate::Todo;
+use mox::mox;
 use moxie_dom::{
     elements::{
         html::*,
