+++ dom/examples/todo/src/footer.rs
@@ -1,4 +1,5 @@
 use crate::{filter::filter, Todo};
+use mox::mox;
 use moxie_dom::{
     elements::{forms::Button, html::*, sectioning::Footer, text_semantics::Span},
     prelude::*,
