+++ dom/examples/todo/src/input.rs
@@ -1,3 +1,4 @@
+use mox::mox;
 use moxie_dom::{
     elements::forms::{input, Input},
     prelude::*,
