+++ dom/examples/todo/src/item.rs
@@ -1,4 +1,5 @@
 use crate::{input::text_input, Todo};
+use mox::mox;
 use moxie_dom::{
     elements::{
         forms::Input,
