+++ dom/examples/todo/src/lib.rs
@@ -1,6 +1,7 @@
 use filter::Visibility;
 use header::input_header;
 use main_section::main_section;
+use mox::mox;
 use moxie_dom::{
     elements::text_content::{div, Div},
     interfaces::node::Node,
